+++ library/src/main/kotlin/dk/ilios/realmfieldnames/FileGenerator.kt
@@ -26,13 +26,7 @@ class FileGenerator(private val filer: Filer) {
      * @return `true` if the files where generated, `false` if not.
      */
     fun generate(fileData: Set<ClassData>): Boolean {
-        for (classData in fileData) {
-            if (!generateFile(classData, fileData)) {
-                return false
-            }
-        }
-
-        return true
+        return fileData.all { generateFile(it, fileData) }
     }
 
     private fun generateFile(classData: ClassData, classPool: Set<ClassData>): Boolean {