+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.1.1
-- Fixed a bug that crashed the annotation processor if Realm wasn't on the classpath (#14)
+- Fixed a bug that crashed the annotation processor if Realm wasn't on the classpath (#14).
+- Fixed crash when processing Kotlin model classes (#9).
 
 ### 1.1.0
  - Fixed a bug where hungarian notation was detected in the middle of a String.